Extend preproccessTag to allow returning an array.
Remove the multiline code. Instead, preprocessTag gets called with the
input and can return an array of tags. In addition, preprocess tags gets called
with whether the event was triggered by a paste or not, as the user may
want different behaviour when trying to paste in text than when typing.
Also allow the type of the input field to be 'email', so on mobile it
displays the appropriate keyboard.